Move settings related content to AutoSettings class
